chore: Update Dockerfile for backend and front-end
This commit updates the Dockerfile for the backend and front-end services. It changes the working directory to the root directory ("/") in the backend Dockerfile and updates the dependencies installation process in both Dockerfiles to install the latest versions of the packages. These changes improve the Docker build process and ensure that the latest dependencies are used.

```