feat: Add user profile data to login flow
This commit adds the retrieval of user profile data to the login flow. After a successful login, the `fetchUserProfile` function is called to make a request to the server and retrieve the user profile data. The received data is then dispatched to the Redux store using the `setUserSuccess` action. This enhancement improves the user experience by providing access to the user's profile information upon login.